build: add Makefile and revamp build script and Dockerfile

+++ build/build.sh
@@ -1,9 +1,35 @@
-#!/bin/bash
-set -eu
+#!/bin/bash -eux
 
-SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )"
+OPTIND=1
 IMAGE_NAME="perlin/wavelet"
-HOST_BUILD_BIN="${SCRIPT_DIR}/_bin"
+SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )"
+HOST_BUILD_BIN="${SCRIPT_DIR}/bin"
+OS_ARCH="linux-amd64"
+CLEAR_BUILDS=false
+
+function show_help {
+    echo "Usage: build.sh [-h] [-d] [-a arch] [-o output]"
+    echo "    -h    Display this help message."
+    echo "    -a    OS and architecture (default: ${OS_ARCH}). You can specify multiple values (e.g. linux-amd64,darwin-amd64,windows-amd64)."
+    echo "    -d    Delete old builds (default: ${CLEAR_BUILDS})."
+    echo "    -o    Binary output directory (default: ${HOST_BUILD_BIN})."
+}
+
+while getopts "h?a:o:d" opt; do
+    case "$opt" in
+    h|\?)
+        show_help
+        exit 0
+        ;;
+    a)  OS_ARCH="$OPTARG"
+        ;;
+    d)  CLEAR_BUILDS=true
+        ;;
+    o)  HOST_BUILD_BIN="$OPTARG"
+        ;;
+    esac
+done
+shift $((OPTIND-1))
 
 cd ${SCRIPT_DIR}/..
 
@@ -11,16 +37,22 @@ cd ${SCRIPT_DIR}/..
 go mod vendor
 
 # clear out old builds
-rm -rf ${HOST_BUILD_BIN}
+if [ ${CLEAR_BUILDS} = true ] && [ -d ${HOST_BUILD_BIN} ]; then
+    rm -rf ${HOST_BUILD_BIN}
+fi
+
+if [ ! -d ${HOST_BUILD_BIN} ]; then
+    mkdir -p ${HOST_BUILD_BIN}
+fi
 
 # make a workspace to compile and test
 docker build \
     --tag ${IMAGE_NAME} \
     --file build/Dockerfile \
+    --build-arg OS_ARCH=${OS_ARCH} \
     $(pwd)
 
 # copy the binaries to the host bin directory
-mkdir -p ${HOST_BUILD_BIN}
 CONTAINER_BUILD_BIN=$(docker run --rm ${IMAGE_NAME} bash -c "echo \$BUILD_BIN")
 docker run \
     --rm \
